log scale first attempt
