+++ samples/line/basic-line.html
@@ -25,7 +25,6 @@
 
 
     <script src="../../dist/apexcharts.js"></script>
-    <script src="../assets/stock-prices.js"></script>
 
     <script>
         var options = {
@@ -44,7 +43,7 @@
             },
             series: [{
                 name: "Desktops",
-                data: [10, 41, 35, 51, 49, 62, 69, 91, 14888]
+                data: [10, 41, 35, 51, 49, 62, 69, 91, 148]
             }],
             title: {
                 text: 'Product Trends by Month',
@@ -56,12 +55,8 @@
                     opacity: 0.5
                 },
             },
-            labels: series.monthDataSeries1.dates,
             xaxis: {
                 categories: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep'],
-            },
-            yaxis: {
-                logarithmic: true
             }
         }
