+++ samples/line/logarithmic-line.html
@@ -1,0 +1,165 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Basic Line Chart</title>
+
+
+    <link href="../assets/styles.css" rel="stylesheet" />
+
+    <style>
+        #chart {
+            max-width: 650px;
+            margin: 35px auto;
+        }
+    </style>
+</head>
+
+<body>
+    <div id="chart">
+
+    </div>
+
+
+    <script src="../../dist/apexcharts.js"></script>
+
+    <script>
+        var options = {
+            chart: {
+                height: 350,
+                type: 'line',
+                zoom: {
+                    enabled: false
+                }
+            },
+            dataLabels: {
+                enabled: false
+            },
+            stroke: {
+                curve: 'straight'
+            },
+            series: [{
+                name: "Desktops",
+                data: [{
+                        x: 1994,
+                        y: 25437639
+                    },
+                    {
+                        x: 1995,
+                        y: 44866595
+                    },
+                    {
+                        x: 1996,
+                        y: 77583866
+                    },
+                    {
+                        x: 1997,
+                        y: 120992212
+                    },
+                    {
+                        x: 1998,
+                        y: 188507628
+                    },
+                    {
+                        x: 1999,
+                        y: 281537652
+                    },
+                    {
+                        x: 2000,
+                        y: 414794957
+                    },
+                    {
+                        x: 2001,
+                        y: 502292245
+                    },
+                    {
+                        x: 2002,
+                        y: 665065014
+                    },
+                    {
+                        x: 2003,
+                        y: 781435983
+                    },
+                    {
+                        x: 2004,
+                        y: 913327771
+                    },
+                    {
+                        x: 2005,
+                        y: 1030101289
+                    },
+                    {
+                        x: 2006,
+                        y: 1162916818
+                    },
+                    {
+                        x: 2007,
+                        y: 1373226988
+                    },
+                    {
+                        x: 2008,
+                        y: 1575067520
+                    },
+                    {
+                        x: 2009,
+                        y: 1766403814
+                    },
+                    {
+                        x: 2010,
+                        y: 2023202974
+                    },
+                    {
+                        x: 2011,
+                        y: 2231957359
+                    },
+                    {
+                        x: 2012,
+                        y: 2494736248
+                    },
+                    {
+                        x: 2013,
+                        y: 2728428107
+                    },
+                    {
+                        x: 2014,
+                        y: 2956385569
+                    },
+                    {
+                        x: 2015,
+                        y: 3185996155
+                    },
+                    {
+                        x: 2016,
+                        y: 3424971237
+                    }
+                ]
+            }],
+            title: {
+                text: 'Product Trends by Month',
+                align: 'left'
+            },
+            grid: {
+                row: {
+                    colors: ['#f3f3f3', 'transparent'], // takes an array which will be repeated on columns
+                    opacity: 0.5
+                },
+            },
+            yaxis: {
+                logarithmic: true
+            }
+        }
+
+        var chart = new ApexCharts(
+            document.querySelector("#chart"),
+            options
+        );
+
+        chart.render();
+
+</script>
+</body>
+
+</html>
