+++ src/charts/Line.js
@@ -1,4 +1,4 @@
-// import Utils from '../utils/Utils'
+import CoreUtils from '../modules/CoreUtils'
 import Graphics from '../modules/Graphics'
 import Fill from '../modules/Fill'
 import DataLabels from '../modules/DataLabels'
@@ -43,7 +43,12 @@ class Line {
       'clip-path': `url(#gridRectMask${w.globals.cuid})`
     })
 
+    const coreUtils = new CoreUtils(this.ctx, w)
+    series = coreUtils.getLogSeries(series)
+
     let yRatio = this.xyRatios.yRatio
+
+    yRatio = coreUtils.getLogYRatios(yRatio)
 
     let zRatio = this.xyRatios.zRatio
     let xRatio = this.xyRatios.xRatio
@@ -196,14 +201,6 @@ class Line {
             y = (zeroY - series[i][j + 1] / yRatio[this.yaxisIndex])
           }
         }
-
-        const logY = () => {
-          const logVal = Math.log(w.globals.maxYArr[this.yaxisIndex]) / Math.log(series[i][j + 1])
-
-          return logVal * yRatio[this.yaxisIndex]
-        }
-
-        y = logY()
 
         // push current X
         xArrj.push(x)
